Pin the JDK used for gradle to JDK 8.
In gradle.properties, we pass the arg 'MaxPermSize' to 'org.gradle.jvmargs'. This property was deprecated and removed in future versions of the JDK, and so without specifying the toolchain to use gradle may default to a later version. In this case, IntelliJ will raise an error when opening the repo 'Unable to start the daemon process'. A number of engineers encountered this issue when working through the 'Contributing a patch to BDP' workshops today.